requirements.txt
google-cloud-aiplatform
langchain-community
langchain-google-vertexai
langchain
gradio
python-telegram-bot
pandas


